+++ src/ipyaladin/widget.py
@@ -25,6 +25,8 @@ from astropy.coordinates.name_resolve import NameResolveError
 from astropy.table.table import QTable, Table
 from astropy.io import fits as astropy_fits
 from astropy.io.fits import HDUList
+from astropy import units as u
+from astropy.units import Quantity
 from astropy.wcs import WCS
 import numpy as np
 import traitlets
@@ -225,14 +227,28 @@ class Aladin(anywidget.AnyWidget):
         # https://github.com/jupyter-widgets/ipywidgets/blob/main/python/ipywidgets/ipywidgets/widgets/domwidget.py
         for key in ["layout", "tabbable", "tooltip"]:
             init_options.pop(key, None)
-        init_options["north_pole_orientation"] = init_options.get(
-            "north_pole_orientation", init_options.get("rotation", 0)
-        )
+
         # some init options are properties here
         self.height = init_options.get("height", self._height)
-        self.rotation = init_options.get("north_pole_orientation", self._rotation)
-        self.target = init_options.get("target", self._target)
-        self.fov = init_options.get("fov", self._fov)
+
+        # make these attrs json serializable before adding to _init_options traitlet
+        # rotation
+        rotation = init_options.pop("rotation", self._rotation)
+        if hasattr(rotation, "unit"):
+            rotation = rotation.to_value("deg")
+        self.rotation = init_options["rotation"] = rotation
+
+        # fov
+        fov = init_options.pop("fov", self._fov)
+        if hasattr(fov, "unit"):
+            fov = fov.to_value("deg")
+        self.fov = init_options["fov"] = fov
+
+        # target
+        self.target = init_options.pop("target", self._target)
+        if isinstance(self.target, SkyCoord):
+            init_options["target"] = self._target
+
         # apply different default options from Aladin-Lite
         ipyaladin_default = {
             "show_simbad_pointer_control": True,
@@ -253,7 +269,7 @@ class Aladin(anywidget.AnyWidget):
         # leading to _is_loaded = True before observing it.
         # If this is the case there should not be any problem because no python commands
         # are expected to be called on an object before the object itself
-        # is instanciated.
+        # is instantiated.
         self.observe(on_load_change, names="_is_loaded")
 
     def _handle_custom_message(self, _: any, message: dict, buffers: any) -> None:
@@ -336,9 +352,9 @@ class Aladin(anywidget.AnyWidget):
         return Angle(self._rotation, unit="deg")
 
     @rotation.setter
-    def rotation(self, rotation: Union[float, Angle]) -> None:
-        if isinstance(rotation, Angle):
-            rotation = rotation.deg
+    def rotation(self, rotation: Union[float, Angle, Quantity]) -> None:
+        if isinstance(rotation, u.Quantity):
+            rotation = rotation.to_value(u.deg)
         if np.isclose(self._rotation, rotation):
             return
         self._rotation = rotation
@@ -414,9 +430,9 @@ class Aladin(anywidget.AnyWidget):
         return Angle(self._fov, unit="deg")
 
     @fov.setter
-    def fov(self, fov: Union[float, Angle]) -> None:
-        if isinstance(fov, Angle):
-            fov = fov.deg
+    def fov(self, fov: Union[float, Angle, Quantity]) -> None:
+        if isinstance(fov, Quantity):
+            fov = fov.to_value(u.deg)
         if np.isclose(fov, self._fov):
             return
         self._fov = fov
@@ -753,8 +769,8 @@ class Aladin(anywidget.AnyWidget):
         """
         warnings.warn(
             "add_moc_from_URL is replaced by add_moc that detects automatically"
-            "that the MOC was given as an URL."
-            "This will be removed in version 1.0.0 (coming after 0.5).",
+            "that the MOC was given as an URL. "
+            "This will be removed in version 1.0.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -780,8 +796,8 @@ class Aladin(anywidget.AnyWidget):
         """
         warnings.warn(
             "add_moc_from_dict is replaced by add_moc that detects automatically"
-            "that the MOC was given as a dictionary."
-            "This will be removed in version 1.0.0 (coming after 0.5).",
+            "that the MOC was given as a dictionary. "
+            "This will be removed in version 1.0.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -958,7 +974,7 @@ class Aladin(anywidget.AnyWidget):
         warnings.warn(
             "'add_overlay_from_stcs' is deprecated, "
             "use 'add_graphic_overlay_from_stcs' instead. "
-            "This will be removed in version 1.0.0 (coming after 0.5).",
+            "This will be removed in version 1.0.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -1039,8 +1055,8 @@ class Aladin(anywidget.AnyWidget):
         This method is deprecated, use selection instead
         """
         warnings.warn(
-            "rectangular_selection is deprecated, use selection('rectangle') instead"
-            "This will be removed in version 1.0.0 (coming after 0.5).",
+            "rectangular_selection is deprecated, use selection('rectangle') instead "
+            "This will be removed in version 1.0.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -1093,8 +1109,8 @@ class Aladin(anywidget.AnyWidget):
 
         """
         warnings.warn(
-            "add_listener is deprecated, use set_listener instead"
-            "This will be removed in version 1.0.0 (coming after 0.5).",
+            "add_listener is deprecated, use set_listener instead. "
+            "This will be removed in version 1.0.0.",
             DeprecationWarning,
             stacklevel=2,
         )